autotest-future: create test specs that read settings from json file (not yet integrated into testers)
