+++ testers/testers/markus_tester.py
@@ -6,128 +6,6 @@ import os
 import subprocess
 import sys
 
-class MarkusTestSpecs(collections.MutableMapping):
-
-    # special keys
-    MATRIX_KEY = 'matrix'
-    MATRIX_NODATA_KEY = 'nodata'
-    MATRIX_NONTEST_KEY = 'extra'
-    FEEDBACK_FILE_KEY = 'feedback_file'
-    TEST_SUITE_KEY = 'test_suite_name'
-    TEST_TIMEOUT_KEY = 'test_timeout'
-    TEST_CASES_KEY = 'test_cases'
-    DATA_FILES_SEPARATOR = ','
-
-    def __init__(self, path_to_specs=None):
-        warn = bool(path_to_specs)
-        if path_to_specs is None: # try to find specs automagically in the virtual environment
-            path_to_specs = sys.executable.replace('venvs', 'specs').replace('bin/python3', 'specs.json')
-        try:
-            with open(path_to_specs) as f:
-                self._specs = json.load(f)
-        except (FileNotFoundError, json.JSONDecodeError) as e:
-            if warn:
-                warnings.warn(str(e))
-            self._specs = {}
-        if MarkusTestSpecs.MATRIX_KEY not in self._specs:
-            self._specs[MarkusTestSpecs.MATRIX_KEY] = {}
-        if MarkusTestSpecs.FEEDBACK_FILE_KEY not in self._specs:
-            self._specs[MarkusTestSpecs.FEEDBACK_FILE_KEY] = None
-        if MarkusTestSpecs.TEST_SUITE_KEY not in self._specs:
-            self._specs[MarkusTestSpecs.TEST_SUITE_KEY] = 'all-tests'
-        if MarkusTestSpecs.TEST_TIMEOUT_KEY not in self._specs:
-            self._specs[MarkusTestSpecs.TEST_TIMEOUT_KEY] = 3600
-        if MarkusTestSpecs.TEST_CASES_KEY not in self._specs:
-            self._specs[MarkusTestSpecs.TEST_CASES_KEY] = 100
-
-    def _setitem(self, key, value):
-        self._specs[key] = value
-
-    def _set_points(self, _, value):
-        """
-        SPECS['points'] = {'test1': {'data1': 11, 'data2': 12}, 'test2': {'data1': 21, 'data2': 22}}
-        Assigns points to the passed tests and datasets, creating them if they don't exist yet.
-        """
-        for test, data_files in value.items():
-            if test in self.matrix:
-                for data_file, points in data_files.items():
-                    self.matrix[test][data_file] = points
-            else:
-                self.matrix[test] = data_files
-
-    def _set_test_points(self, _, value):
-        """
-        SPECS['test_points'] = {'test1': 1, 'test2': 2}
-        Assigns points to all datasets of the passed tests, creating the tests if they don't exist yet.
-        """
-        for test, points in value.items():
-            data_files = self.matrix.setdefault(test, {MarkusTestSpecs.MATRIX_NODATA_KEY: {}})
-            for data_file in data_files:
-                if data_file == MarkusTestSpecs.MATRIX_NONTEST_KEY:
-                    continue
-                self.matrix[test][data_file] = points
-
-    def _set_data_points(self, _, value):
-        """
-        SPECS['data_points'] = {'data1': 1, 'data2': 2}
-        Assigns points to all existing tests that use the passed datasets, does nothing for tests that don't use them.
-        """
-        for test in self.matrix:
-            for data_file, points in value.items():
-                if data_file not in self.matrix[test]:
-                    continue
-                self.matrix[test][data_file] = points
-
-    def _set_all_points(self, _, points):
-        """
-        SPECS['all_points'] = points
-        Assigns points to all existing tests and datasets.
-        """
-        for test, data_files in self.matrix.items():
-            for data_file in data_files:
-                if data_file == MarkusTestSpecs.MATRIX_NONTEST_KEY:
-                    continue
-                self.matrix[test][data_file] = points
-
-    def _get_test_suite_name(self, key):
-        name = self._specs['test_suite_name']
-        if isinstance(name, str):
-            return {t:name for t in self.tests}
-        return name
-
-    def __setitem__(self, key, value):
-        switch = {'points': self._set_points, 'test_points': self._set_test_points,
-                  'data_points': self._set_data_points, 'all_points': self._set_all_points}
-        setter = switch.get(key, self._setitem)
-        setter(key, value)
-
-    def __getitem__(self, key):
-        switch = {'test_suite_name': self._get_test_suite_name}
-        getter = switch.get(key, self._specs.__getitem__)
-        return getter(key)
-
-    def __delitem__(self, key):
-        del self._specs[key]
-
-    def __iter__(self):
-        return iter(self._specs)
-
-    def __len__(self):
-        return len(self._specs)
-
-    @property
-    def matrix(self):
-        return self[MarkusTestSpecs.MATRIX_KEY]
-
-    @property
-    def feedback_file(self):
-        return self[MarkusTestSpecs.FEEDBACK_FILE_KEY]
-
-    @property
-    def tests(self):
-        return self.matrix.keys()
-
-
 class MarkusTest:
 
     class Status(enum.Enum):
